Changed libx264-5.json, further test needed
I've used VSCode and edited those 2 files, and made sure the json format was correct and all parameters are linked to the definitions in the bottom part of file.
However due to I've mainly worked with x265.exe instead of libx265, some parameters are missing in libx265 and I cannot gurantee they actually will work (some parameter may get deleted or the definitions needs to be modified if libx265 v3.5 persists not recognizing)
Software is not my profession and I wish I could compile and test the changes myself. Someone needs to compile and check in software to see if the parameters are applied correctly.
Changes also happened in x264 parameters, they have been separated with color profiles (4:2:0 and 4:4:4) thus creates more versions, just 1 version needs to be tested for them.